Rename events.json to spot_events.json for clarity
This JSON fixture is an example of the data the API provides when
fetching recent checkin events for a particular spot
(e.g., /spots/452593/events).  It includes checkins for many
different users at the same spot.

This data is different than the event data the API provides when
fetching recent checkin events for a particular user
(e.g., /users/1707/events).  That data would include many checkins at
different spots for the same user.